% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-pvalues.R
\name{calc_prob_null2}
\alias{calc_prob_null2}
\title{Calculate probability for given parameters}
\usage{
calc_prob_null2(
  theta_cands,
  psi,
  psi0,
  minus1,
  SSpacearr,
  logC,
  II,
  psi_v = FALSE
)
}
\arguments{
\item{theta_cands}{A matrix with samples in the rows and the parameters in
the columns}

\item{psi}{The function of interest mapping parameters to the real line}

\item{psi0}{The null boundary for testing psi <= psi0}

\item{minus1}{Either plus or minus 1}

\item{SSpacearr}{A matrix with the sample space for the given size of the
problem}

\item{logC}{log multinomial coefficient for each element of the sample space}

\item{II}{logical vector of sample space psi being more extreme than the
observed psi}

\item{psi_v}{Is psi vectorized by row?}
}
\value{
A numeric vector of probabilities
}
\description{
Given a set of candidate parameter vectors, check if the null \eqn{\psi \leq
\psi_0} is satisfied, and if so, compute the probability for each element of
the sample space
}
