% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinate}
\alias{combinate}
\title{Calculate combinations of multinomials}
\usage{
combinate(X, Y)
}
\arguments{
\item{X}{Matrix 1}

\item{Y}{Matrix 2}
}
\value{
A list of arrays
}
\description{
Calculate combinations of multinomials
}
