% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_index}
\alias{expand_index}
\title{Get a matrix of indices for all possible combinations of vectors of lengths}
\usage{
expand_index(lengths)
}
\arguments{
\item{lengths}{A vector with the lengths of each index to expand}
}
\value{
A matrix with length(lengths) columns and prod(lengths) rows
}
\description{
This is basically the same as \link[base]{expand.grid}, but faster for integers
}
\examples{
expand_index(c(2, 3, 4))
## the same as
expand.grid(1:2, 1:3, 1:4)

}
