% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theta_random}
\alias{get_theta_random}
\title{Sample from the unit simplex in d dimensions}
\usage{
get_theta_random(d = 4, nsamp = 75)
}
\arguments{
\item{d}{the dimension}

\item{nsamp}{the number of samples to take uniformly in the d space}
}
\value{
The grid over Theta, the parameter space. A matrix with d columns and nsamp rows
}
\description{
Sample from the unit simplex in d dimensions
}
\examples{
get_theta_random(3, 10)

}
