% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_multinom_coef}
\alias{log_multinom_coef}
\title{Calculate log of multinomial coefficient}
\usage{
log_multinom_coef(x, sumx)
}
\arguments{
\item{x}{Vector of observed counts in each cell}

\item{size}{Total count}
}
\value{
The log multinomial coefficient
}
\description{
Calculate log of multinomial coefficient
}
\examples{
#' @examples
S0 <- sspace_multinom(4, 6)
S1 <- sspace_multinom(4, 7)
logC0<- apply(S0,1,log_multinom_coef,sumx=6)
logC1<- apply(S1,1,log_multinom_coef,sumx=7)
logC<- outer(logC0,logC1,'+')

}
