% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xactonomial.R
\name{pvalue_psi0}
\alias{pvalue_psi0}
\title{Compute a p value for the test of psi <= psi0 (lower = TRUE) or psi >= psi0 (lower = FALSE)}
\usage{
pvalue_psi0(
  psi0,
  psi,
  psi_hat,
  psi_obs,
  alternative = "two.sided",
  maxit,
  chunksize,
  p_target,
  SSpacearr,
  logC,
  d_k,
  psi_is_vectorized = FALSE,
  theta_sampler = runif_dk_vects,
  ga = FALSE,
  ga_gfactor = 1,
  ga_lrate = 0.01,
  ga_restart_every = 10
)
}
\arguments{
\item{psi0}{The null value}

\item{psi}{The function of interest}

\item{psi_hat}{The vector of psi values at each element of the sample space}

\item{psi_obs}{The observed estimate}

\item{maxit}{Maximum iterations}

\item{chunksize}{Chunk size}

\item{SSpacearr}{The sample space array}

\item{logC}{The log multinomial coefficient}

\item{d_k}{The vector of dimensions}

\item{theta_sampler}{Sampler}

\item{lower}{Do a one sided test of the null that it is less than psi0, otherwise greater.}

\item{target}{Stop the algorithm if p >= target (for speed)}

\item{psi_v}{Is psi vectorized by row?}

\item{gradient_ascent}{Do ga?}

\item{gamma}{Parameter to determine concentration of dirichlet, only relevant if gradient_ascent = TRUE}
}
\value{
A p-value
}
\description{
Compute a p value for the test of psi <= psi0 (lower = TRUE) or psi >= psi0 (lower = FALSE)
}
