% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirich_dk_vects}
\alias{rdirich_dk_vects}
\title{Sample independently from Dirichlet distributions for each of d_k vectors}
\usage{
rdirich_dk_vects(nsamp, alpha)
}
\arguments{
\item{nsamp}{number of samples to take}

\item{alpha}{List of vectors of concentration parameters}
}
\value{
A matrix with sum(d_k) columns and nsamp rows
}
\description{
Sample independently from Dirichlet distributions for each of d_k vectors
}
\examples{
rdirich_dk_vects(10, list(rep(1, 3), rep(1, 4), rep(1, 2)))

}
