% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_sspace}
\alias{rust_sspace}
\title{calculate sample space of a multinomial with dimension d and sample size n}
\usage{
rust_sspace(d, n)
}
\description{
calculate sample space of a multinomial with dimension d and sample size n
}
