% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sample_unit_simplex2}
\alias{sample_unit_simplex2}
\title{Return a random sample from the d unit simplex}
\usage{
sample_unit_simplex2(d)
}
\description{
Return a random sample from the d unit simplex
}
