% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sample_unit_simplexn}
\alias{sample_unit_simplexn}
\title{Return n random samples from the d unit simplex}
\usage{
sample_unit_simplexn(d, n)
}
\description{
Return n random samples from the d unit simplex
}
