% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{xactonomial-package}
\alias{xactonomial-package}
\title{xactonomial: Exact Inference for Real-Valued Functionals of k-Sample Multinomial Parameters}
\description{
We consider the k sample multinomial problem where we observe k vectors (possibly of different lengths), each representing an independent sample from a multinomial. For a given function psi which takes in the concatenated vector of multinomial probabilities and outputs a real number, we are interested in constructing a confidence interval for psi. We use an exact (but computational and stochastic) method to compute a confidence interval and a function for calculation of p values. The details of the method will be in a forthcoming preprint.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sachsmc/xactonomial}
  \item Report bugs at \url{https://github.com/sachsmc/xactonomial/issues}
}

}
\author{
\strong{Maintainer}: Michael C Sachs \email{sachsmc@gmail.com}

Authors:
\itemize{
  \item Michael P Fay
}

}
