% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xactonomial.R
\name{xactonomial}
\alias{xactonomial}
\title{Exact inference for a real-valued function of multinomial parameters}
\usage{
xactonomial(psi, data, alpha = 0.05, psi_limits, maxit = 50, chunksize = 500)
}
\arguments{
\item{psi}{Function that takes in a vector of parameters and outputs a real
valued number}

\item{data}{A list with k elements representing the vectors of counts of a
k-sample multinomial}

\item{alpha}{A 1 - alpha percent confidence interval will be computed}

\item{psi_limits}{A vector of length 2 giving the lower and upper limits of
the range of \eqn{\psi(\theta)}}

\item{maxit}{Maximum number of iterations of the stochastic procedure}

\item{chunksize}{The number of samples taken from the parameter space at each
iteration}
}
\value{
A list with 3 elements: the estimate, the 1 - alpha percent
  confidence interval, and a function for calculation of p-values
}
\description{
We consider the k sample multinomial problem where we observe k vectors
(possibly of different lengths), each representing an independent sample from
a multinomial. For a given function psi which takes in the concatenated
vector of multinomial probabilities and outputs a real number, we are
interested in constructing a confidence interval for psi.
}
\details{
Let \eqn{T_j} be distributed
\eqn{\mbox{Multinomial}_{d_j}(\boldsymbol{\theta}_j, n_j)} for \eqn{j = 1,
\ldots, k} and denote \eqn{\boldsymbol{T} = (T_1, \ldots, T_k)} and
\eqn{\boldsymbol{\theta} = (\theta_1, \ldots, \theta_k)}. The subscript
\eqn{d_j} denotes the dimension of the multinomial. Suppose one is interested
in the parameter \eqn{\psi(\boldsymbol{\theta}) \in \Theta \subseteq
\mathbb{R}}. Given a sample of size \eqn{n} from \eqn{\boldsymbol{T}}, one
can estimate \eqn{\boldsymbol{\theta}} with the sample proportions as
\eqn{\hat{\boldsymbol{\theta}}} and hence
\eqn{\pi(\hat{\boldsymbol{\theta}})}. This function constructs a \eqn{1 -
\alpha} percent confidence interval for \eqn{\psi(\boldsymbol{\theta})} and
provides a function to calculate a p value for a test of the null hypothesis
\eqn{H_0: \psi(\boldsymbol{\theta}) \leq \psi_0}. We make no assumptions and
do not rely on large sample approximations. The computation in somewhat
involved so it is best for small sample sizes.
}
\examples{
psi_ba <- function(theta) {
  theta1 <- theta[1:4]
  theta2 <- theta[5:8]
  sum(sqrt(theta1 * theta2))
  }
data <- list(T1 = c(2,1,2,1), T2 = c(0,1,3,3))
xactonomial(psi_ba, data, psi_limits = c(0, 1), maxit = 5, chunksize = 20)

}
