% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinate}
\alias{combinate}
\title{Arrange all combinations of rows of two matrices}
\usage{
combinate(X, Y)
}
\arguments{
\item{X}{Matrix 1}

\item{Y}{Matrix 2}
}
\value{
A list containing Sspace, the sample space (vectors of counts), and
 logC, a vector of the log multinomial coefficients.
}
\description{
Given X and Y, both matrices where the rows are counts of multinomial trials,
produce all combinations rowwise, concatenate the rows into a new matrix, and
calculate the log multinomial coefficients for the combination.
}
\examples{
slist_2_3 <- combinate(matrix(sspace_multinom(2, 5), ncol = 2, byrow = TRUE),
   matrix(sspace_multinom(3, 6), ncol = 3, byrow = TRUE))

}
