% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinate2}
\alias{combinate2}
\title{Like combinate but add to existing}
\usage{
combinate2(X, Y)
}
\arguments{
\item{X}{A list as returned by combinate}

\item{Y}{Matrix 2}
}
\value{
A list of arrays
}
\description{
Like combinate but add to existing
}
