% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_multinom_coef}
\alias{log_multinom_coef}
\title{Calculate log of multinomial coefficient}
\usage{
log_multinom_coef(x, sumx)
}
\arguments{
\item{x}{Vector of observed counts in each cell}

\item{sumx}{Total count}
}
\value{
The vector of log multinomial coefficients
}
\description{
Calculate log of multinomial coefficient
}
\examples{
S0 <- matrix(sspace_multinom(4, 6), ncol = 4, byrow = TRUE)
logC0<- apply(S0,1,log_multinom_coef,sumx=6)

}
