% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runif_dk_vects}
\alias{runif_dk_vects}
\title{Sample uniformly and independently from d_k simplices}
\usage{
runif_dk_vects(d_k, nsamp)
}
\arguments{
\item{d_k}{vector of vector lengths}

\item{nsamp}{number of samples to take}
}
\value{
A matrix with sum(d_k) columns and nsamp rows
}
\description{
Sample uniformly and independently from d_k simplices
}
\examples{
runif_dk_vects(c(3, 4, 2), 10)
}
