% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sample_unit_simplexn}
\alias{sample_unit_simplexn}
\title{Sample n times from the unit simplex in d dimensions}
\usage{
sample_unit_simplexn(d, n)
}
\arguments{
\item{d}{the dimension}

\item{n}{the number of samples to take uniformly in the d space}
}
\value{
The grid over Theta, the parameter space. To be converted to a matrix with d columns and nsamp rows
}
\description{
Sample n times from the unit simplex in d dimensions
}
\examples{
matrix(sample_unit_simplexn(3, 10), ncol = d, byrow = TRUE)
}
