% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sspace_multinom_sample}
\alias{sspace_multinom_sample}
\title{Get a sumsample of the multinomial sample space}
\usage{
sspace_multinom_sample(d, n, k)
}
\arguments{
\item{d}{The dimension}

\item{n}{The sample size}

\item{k}{The number of elements to keep}
}
\value{
A vector with a subsample from the sample space, to be converted to a matrix
with d columns and k <= choose(n + d - 1, d - 1) rows
}
\description{
Get a sumsample of the multinomial sample space
}
\examples{
matrix(sspace_multinom_sample(3, 5, 10), ncol = 3, byrow = TRUE)
}
