% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{xactonomial-package}
\alias{xactonomial-package}
\title{xactonomial: Exact Inference for Real-Valued Functionals of k-Sample Multinomial Parameters}
\description{
We consider the k sample multinomial problem where we observe k vectors (possibly of different lengths), each representing an independent multinomial random vector. For a given function psi which takes in the concatenated vector of multinomial probabilities and outputs a real number, this is a Monte Carlo estimation procedure of an exact p-value and confidence interval. The resulting inference is valid even in small samples, and when the parameter is on the boundary, and when the function is nondifferentiable at the parameter value, all situations where asymptotic methods and the bootstrap would fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sachsmc/xactonomial}
  \item Report bugs at \url{https://github.com/sachsmc/xactonomial/issues}
}

}
\author{
\strong{Maintainer}: Michael C Sachs \email{sachsmc@gmail.com}

Authors:
\itemize{
  \item Michael P Fay
}

}
